requirements.txt
langchain==0.3.0
langchain-community==0.3.0
streamlit==1.38.0
fpdf2==2.7.9
GitPython==3.1.43

-e .